Move all legacy endpoints into legacy_routes.py
