+++ sogs/__init__.py
@@ -2,6 +2,7 @@
 
 from . import routes
 from . import onion_request
+from . import legacy_routes
 from .web import app
 
 import coloredlogs
